Compatibility changes for experiment build

+++ sar_control/include/shared_lib.h
@@ -18,8 +18,8 @@ extern "C" {
 #define PWM_MAX 60000
 #define g2Newton (9.81f/1000.0f)
 #define Newton2g (1000.0f/9.81f)
-#define Deg2Rad M_PI/180.0f
-#define Rad2Deg 180.0f/M_PI
+#define Deg2Rad (float)M_PI/180.0f
+#define Rad2Deg 180.0f/(float)M_PI
